Return updated twin version from reported property update call (#2711)
The service always returned this new version after a successful reported properties update, but the SDK wasn't bubbling it up to the user in v1

+++ iothub/device/src/Transport/Mqtt/PatchTwinResponse.cs
@@ -19,7 +19,7 @@ namespace Microsoft.Azure.Devices.Client.Transport.Mqtt
         /// <summary>
         /// The new version of the twin after the patch.
         /// </summary>
-        internal int Version { get; set; }
+        internal long Version { get; set; }
 
         /// <summary>
         /// The error message if the request failed.
